refactor(Volt): RoadRunner integrations
